chore(hardening): close the Free engine at the zero quality gate
Backlog closure sweep, the full docs pass, and the final static-analysis ramp.

Quality (the contractual ZERO gate):
- PHPStan level 6 -> 8 with the baseline kept empty; every source file passes L8.
- ECS skip list emptied (Typesense.php, config.php, the synonyms migration
  cleaned to house standard); ECS clean across src and tests.
- Deleted the dead Vue-era control-panel apparatus that was holding untyped,
  orphaned code: the src/web/assets Vue/TS sources and dist output, the
  TypesenseAsset and TypesenseUtilityAsset bundles, and the Vite plugin service
  wiring in PluginTrait/TypesenseVariable. Nothing rendered them after the
  Collections and Synonyms screens were retired.
- Rewrote PluginTrait to house standard with typed, documented component getters.
- Modernised the shipped src/config.php from the legacy shim shape to the fluent
  builder (null-safe transform).

Regression tests (the three damning themes + one):
- #68 CP action URLs are CP-based, never the site URL (unit, UrlHelper::actionUrl).
- #63 saving before configuration is a safe no-op.
- #67 applying a draft skips cleanly instead of fataling.
- #65/#66 one element indexes into every collection whose query matches it.

Docs:
- Completed docs/ for the whole Free surface: multisite, drift, backup, and
  extension events for plugin authors. Rewrote the README for 5.9.0 (Free/Pro
  split, server floor v28 and refused 30.0/30.1, quick start).

Backlog sweep recorded in smoke-testing/p7-issue-sweep.md (local, nothing posted):
7 issues and 8 PRs dispositioned, 6 pinned by tests. Search bug fixed along the
way: Search now resolves the site-aware target so collectionPerSite front-end
searches hit the right collection.

+++ src/services/Drift.php
@@ -119,7 +119,7 @@ class Drift extends Component
             return $targets;
         }
 
-        $primaryId = Craft::$app->getSites()->getPrimarySite()->id;
+        $primaryId = (int)Craft::$app->getSites()->getPrimarySite()->id;
 
         return [$primaryId => $registry->resolveName($collection, $primaryId)];
     }
